moved font-loader to live with typography
