+++ layout/wrapper.html
@@ -74,7 +74,7 @@
 				{{#if anon}}
 				window.nextIsAnonymous = {{anon.userIsAnonymous}};
 				{{/if}}
-				{{>n-ui/_initlayout/partialslates/bootstrapper}}
+				{{>n-ui/layout/partials/bootstrapper}}
 			</script>
 			{{/if}}
